Can now simplify by removing HtmlComponentBase and changing HtmlComponent to a readonly struct
